r3565| New script function getpixel(handle, x, y, type). Get a pixel from specific screen or sprite.
type - screen or sprite (or other format if supported in the future)
0 - screen
1 - sprite

x,y - position, for sprite, centerx and centery is not considered. If the point falls out of the dimension, the function returns 0.

handle - screen or sprite handle, must be valid.

return - For 8 bit image, it returns the index. For 16 and 32 bit, it returns the color as an integer.