r3972| Don't try to read data folder from the filesystem on Wii.
The only way it would work at a decent speed is with an external USB HD, and
the most common case is running OpenBOR from SD, where the current directory
is the SD root.  Since other Wii homebrew applications use a folder called
"data" on the SD root, this can disable file caching, which results in mods
taking 30+ minutes to load that should load in a few seconds.