fix bug using non-existing "time" variable
this was overlooked in da992740367ce36a49c42ed9121777bd9fa246a6 due to the
definition of the variable inside that function.
this had the effect that the linker found the libc function "time" and linked
it to this spot, therefore no linker error happened.
this was discovered due to ubuntu 20.04 defaulting to having libc compiled as
pie, which made the relocation go boom.